+++ basalt/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [Refactor markdown event parsing](https://github.com/erikjuhani/basalt/commit/7115dfa48368f55a12e3a359c1941026ab203933)
+- [Support loosely checked task items](https://github.com/erikjuhani/basalt/commit/7e14b39fc2b14942ea873377e591adf06cf261bf)
 
 ## 0.3.1